+++ src/Connection/Protocols/Protocol.php
@@ -191,7 +191,6 @@ abstract class Protocol implements ProtocolInterface {
             STREAM_CLIENT_CONNECT,
             stream_context_create($this->defaultSocketOptions($transport))
         );
-        stream_set_timeout($stream, $timeout);
 
         if (!$stream) {
             throw new ConnectionFailedException($errstr, $errno);