+++ CHANGELOG.md
@@ -18,7 +18,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - NaN
 
 ## [2.7.2] - 2021-09-27
-### Added
+### Fixed
 - Fixed problem with skipping last line of the response. #166 (thanks @szymekjanaczek)
 
 ## [2.7.1] - 2021-09-08